+++ test/unit/range/views/CMakeLists.txt
@@ -5,7 +5,6 @@ seqan3_test(as_const_test.cpp)
 seqan3_test(char_to_test.cpp)
 seqan3_test(convert_test.cpp)
 seqan3_test(deep_test.cpp)
-seqan3_test(pairwise_combine_test.cpp)
 seqan3_test(drop_test.cpp)
 seqan3_test(istreambuf_test.cpp)
 seqan3_test(move_test.cpp)
