+++ test/unit/utility/views/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectories()
 seqan3_test (elements_test.cpp)
 seqan3_test (enforce_random_access_test.cpp)
 seqan3_test(interleave_test.cpp)
+seqan3_test(pairwise_combine_test.cpp)
 seqan3_test (repeat_n_test.cpp)
 seqan3_test (repeat_test.cpp)
 seqan3_test (single_pass_input_test.cpp)
